requirements.txt
homeassistant>=2024.6.0
paho-mqtt>=1.6.1
protobuf>=4.25.0